put WinSequence into seperate class
